move calculation of dayInUniversity to tools
